Add munit.internal.io package
