+++ munit/js/src/main/scala/munit/internal/io/File.scala
@@ -1,4 +1,4 @@
-package munit.internal
+package munit.internal.io
 
 import java.net.URI
