+++ munit/js/src/main/scala/munit/internal/io/JSIO.scala
@@ -1,4 +1,4 @@
-package munit.internal
+package munit.internal.io
 
 import scala.scalajs.js
 import scala.util.Try
