+++ munit/js/src/main/scala/munit/internal/io/Paths.scala
@@ -1,4 +1,4 @@
-package munit.internal
+package munit.internal.io
 
 import java.net.URI
